feat(ref-klv-04)!: widen ST 0903 VMTI targetId + pixel fields to u64
VTargetPack target_id + 5 pixel fields (centroid_pixel, bbox_top_left_pixel,
bbox_bottom_right_pixel, centroid_pix_row, centroid_pix_col) -> u64. BER-OID
up to 9 wire bytes per ST 0903.6 §10.2.2.1; var-uint up to 6 bytes for V6.
KlvEncodeError VTargetPackEmpty/DuplicateTargetId target_id -> u64 in lockstep.
encode: buf [0u8;5]->[0u8;10] (u64 targetId needs up to 10 BER-OID bytes);
pixel emits -> emit_var_u64; encoded_len -> ber_oid_len_u64 + var_u64_len.
decode: leading targetId -> read_ber_oid_u64; pixel fields -> read_var_u64;
non-pixel var fields (intensity/algorithm_id/history) keep read_var_u32.
var_uint: remove #[allow(dead_code)] from u64 helpers now in use; fix stale
module-doc "1..=4 byte" note to mention u64 variants accept 1..=8. Small
legacy values remain byte-identical. BREAKING: re-baseline tst-core public-api
(non_exhaustive 269 unchanged).

+++ crates/tst-core/src/error.rs
@@ -165,14 +165,14 @@ pub enum KlvEncodeError {
     #[error(
         "ST 0903 VTarget Pack {target_id} has no TLV items after targetId (ST 0903.4-10 requires at least one)"
     )]
-    VTargetPackEmpty { target_id: u32 },
+    VTargetPackEmpty { target_id: u64 },
 
     /// An ST 0903 vTargetSeries contains more than one VTarget Pack
     /// with the same targetId (ST 0903.6-126 requires unique targetIds).
     #[error(
         "ST 0903 vTargetSeries contains duplicate targetId {target_id} (ST 0903.6-126 requires unique targetIds)"
     )]
-    DuplicateTargetId { target_id: u32 },
+    DuplicateTargetId { target_id: u64 },
 
     /// A standalone ST 0903 VMTI Local Set contains a VTarget Pack
     /// with a parent-relative offset tag, which is forbidden in the
